Rename WECleaner.ps1 to Local-WECleaner.ps1
Run locally on a WEC.

This PowerShell script automates the process of checking and cleaning up outdated computers from Windows Event Collector (WEC) subscriptions. It retrieves a list of subscriptions, iterates through each, and identifies computers with a LastHeartbeatTime older than 7 days. The script then removes these outdated computers from the registry and logs their information in a subscription-specific log file located in the same directory as the script.